+++ packages/manager/src/components/EntityHeader/EntityHeader.stories.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { Box } from 'src/components/Box';
 import { EntityHeader } from 'src/components/EntityHeader/EntityHeader';
 import { Hidden } from 'src/components/Hidden';
-import { LinodeActionMenu } from 'src/features/Linodes/LinodesLanding/LinodeActionMenu';
+import { LinodeActionMenu } from 'src/features/Linodes/LinodesLanding/LinodeActionMenu/LinodeActionMenu';
 
 import { Button } from '../Button/Button';
 import { Link } from '../Link';
