+++ packages/manager/src/features/Account/utils.ts
@@ -8,10 +8,22 @@ import type {
 } from '@linode/api-v4';
 import type { GrantTypeMap } from 'src/features/Account/types';
 
-type ActionType = 'create' | 'delete' | 'edit' | 'view';
+export type ActionType =
+  | 'clone'
+  | 'create'
+  | 'delete'
+  | 'edit'
+  | 'migrate'
+  | 'modify'
+  | 'reboot'
+  | 'rebuild'
+  | 'rescue'
+  | 'resize'
+  | 'view';
 
 interface GetRestrictedResourceText {
   action?: ActionType;
+  includeContactMessage?: boolean;
   isSingular?: boolean;
   resourceType: GrantTypeMap;
 }
@@ -36,6 +48,7 @@ export type RestrictedGlobalGrantType =
  */
 export const getRestrictedResourceText = ({
   action = 'edit',
+  includeContactMessage = true,
   isSingular = true,
   resourceType,
 }: GetRestrictedResourceText): string => {
@@ -43,7 +56,14 @@ export const getRestrictedResourceText = ({
     ? 'this ' + resourceType.replace(/s$/, '')
     : resourceType;
 
-  return `You don't have permissions to ${action} ${resource}. Please contact your account administrator to request the necessary permissions.`;
+  let message = `You don't have permissions to ${action} ${resource}.`;
+
+  if (includeContactMessage) {
+    message +=
+      ' Please contact your account administrator to request the necessary permissions.';
+  }
+
+  return message;
 };
 
 /**
